requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
smartapi-python==1.5.5
pyotp==2.9.0
python-dotenv==1.0.1
